Move Multivariate
move the multi tests to stats
